django_shop/requirements.txt
asn1crypto==0.24.0
Babel==2.7.0
cffi==1.12.3
configparser==3.5.0
cryptography==2.7
Django==2.2.4
django-extensions==2.2.1
django-phonenumber-field==3.0.1
django-split-settings==0.3.0
django-ssl-redirect==2.1.1
django-sslserver==0.21
djangosecure==0.0.11
future==0.16.0
gunicorn==19.9.0
phonenumbers==8.10.17
Pillow==6.1.0
pycparser==2.19
pycryptodome==3.6.1
pyOpenSSL==19.0.0
